[EPGSearch] Fix hang on zapping
If config.usage.multibouquet.value is False, and you enter EPGSearch
and press OK to zap to a channel (including to the current channel),
the UI hangs.

The cause is that the EPGSearch.zap() code assumes that the root bouquet
is a list of bouquets, not a list of channels (which it is if
config.usage.multibouquet.value is False) and the exit conditions for the

while True:

loop are never met.

The fix abstracts the service list search into a local function,
and searches the root bouquet as a list of services (rather than a
list of bouquets) if config.usage.multibouquet.value is False.

Because servicerefs extracted from the EPG don't (and can't) have
valid channel numbers, the new serviceInBouquet() function sets the
channel number of self.currentService.ref from the matched bouquet
entry.

Also, the original test for whether a serviceref is a bouquet was:

(bouquet.flags and int(bouquet.flags) != 519) & eServiceReference.isDirectory

This makes little sense and only works at all because the boolean
result of (bouquet.flags and int(bouquet.flags) != 519), which will
be true for all sorts of serviceref flag values will map to 1 if
True, and then the "&" with eServiceReference.isDirectory will
succeed because eServiceReference.isDirectory happens to have the
value 1.

The actual intent of the test wasn't all that clear, so I have changed
it to:

bouquet.flags & (eServiceReference.isDirectory | eServiceReference.isInvisible) == eServiceReference.isDirectory.