[EPGSearch] Prevent unnecessary bouquet change on zapping
If config.usage.multibouquet.value is True, and you enter EPGSearch
and press OK to zap to a channel (including to the current channel),
the current bouquet will be changed to the first bouquet in the
bouquet list that contains the service, even if the current bouquet
also contains the service.

The fix searches for the service in the current bouquet first, and
only if that fails, searches the other bouquets to match the service
to a bouquet.