EPG Search: Allow specification of search scope
Allow specification of the search scope as all services/all
bouquets/current bouquet/current service.

On Beyonwiz firmware the default search scope is "current bouquet",
and on all firmware, it is "all services".

Adding "ask user" will be in a following commit.

[EPGSearch]

Split searchEPG() into two parts in preparation for the addition
of an "ask user" search mode.

Add search filtering to doEPGSearch(), the part of the old searchEPG()
that execues the search & add methods to build the search filter
sets.

Add a callback to the open of EPGSearchSetup to refresh the list
if the search mode changes.

Move import of ChannelSelection to the global level because it's
now used more than once.

Make bouquets list serviceref string a global variable, because it
is also used more than once.

[EPGSearchSetup]

Add a ConfigListEntry for the new search scope setting.

[__init__]

Add config.plugins.epgsearch.scope and its default setting logic.