EPG Search: Allow "ask user" specification of search scope
Add "ask user" search scope, to allow the user to specify the
scope on every search.

[EPGSearch]

Add the "ask user" popup and the logic to pre-select the
config.plugins.epgsearch.defaultscope setting when it is popped up.

Move the initial search out of onCreate() and into new method
startUp(), because this is too early in the screen setup to invoke
the "ask user" popup.

Add a timer to invoke the first search via startUp() after the
screen setup has completed.  The onShown callback can't be used
because it's called every time the screen is (re)shown, not just
on the first show. The onFirstExecBegin callbacks, like onLayoutFinish
(from which onCreate() is called), are called too early.

Flag whether the search is the first search in the screen, and exit
EPGSearch without saving the search if the "ask user" popup is
cancelled, and simply return to the old search results if the popup
is canceled on subsequent searches in the same screen invocation.

Save the "ask user" choice and use it in refreshlist() if a choice
has been made. This stops the user being asked to choose the search scope
on return from non-search screens, like Add Timer.

Only call refreshlist() in the EPGSearchSetup callback if the new
search scope is not "ask user" and there is no saved "ask user" choice.
This prevents a distracting popup on return from the setup screen.

Tidy up refreshTimer setup to use "callback" instead of "timeout.get()".

[EPGSearchSetup]

Add a ConfigListEntry for the new default search scope setting.
This entry is always shown. It should only be shown when
config.plugins.epgsearch.scope is "ask user". This will be added
in a following commit.

[__init__]

Add config.plugins.epgsearch.defaultscope and its default setting logic.