[EPGSearchSetup] Only show "ask user" specification of search scope when active
Split creation of the config list out into a method, and make the
"Default scope when asked" setting only appear when "Search scope" is set
to "add user".

Add a method to add a notifier on config.plugins.epgsearch.scope and
one to remove the notifier called from onClose. Add a method to be called
from the notifier to reload the config list.