[EPGSearch]: Add "Setup" entry to "ask user" menu
Add a "Setup" entry to the "ask user" menu, so that if the Search
scope is "ask user", the Setup Screen can be entered to change
search settings before the search is done.

When the "Setup" screen exits, return to the "ask user" menu if the
"Setup" screen was called from the "ask user" menu.

Make EPGSearch.firstSearch a property that tests whether "searchargs"
is an attribute.

Move deletion of the "searchargs" attribute until the first search
is actually executed, rather than assuming in EPGSearch.startUp()
that the search will be executed immediately.

Similarly, move the setting of EPGSearch.currSearch to when the search
is actually executed.

Move the handling of the initial searchargs into EPGSearch.refreshlist()
and call it from EPGSearch.startUp().