[EPGSearchList]: Tidy up and remove near-duplicate code
Refactor timer pixmap code so that all pixmaps are handled in a
loop driven by a list of pixmaps to display, instead of a long
if/elif/else chain.

Adjust the width of the title text so that it doesn't overlap with
the timer pixmaps.

Allow the text to use the whole of the remaining space in a row.
The previous implementation used the title width from EPGList which
did not make use of all available space in some implementations.

Allow the display of the AutoTimer pixmap if the EPGList implementation
supports showing it.

Remove unused rec2 local variable.