[EPGSearch]: Add search type/case sensitivity settings
[EPGSearch]

Add code to set the eEPGCache.search() search type and search case
sensitivity from settings.

Generalise the mechanism for detecting setting changes that need
the search to be re-done, and save its state in a private attribute.

Tidy up and correct the logic for testing whether a list re-search
is needed.

[__init__]

Add config variables for search type and case sensitivity.

[EPGSearchSetup]

Add setup entries for search type and case sensitivity.