[epgsearch] Make searches more comprehensive and accurate
Search the EPG on serviceref and then filter on the search string to
preserve path and name information in the servicerefs searched on.

Path and name information is lost when eEPGCache::search() is
used for the search.

Change the "all services" search so that it searches all the
servicerefs in "all bouquets" as well as the servicerefs in lamedb.